Final Project | Payment And Transfer are DONE

+++ Final-BankingSystem/BankingSystem/src/main/java/com/example/bankingsystem/service/impl/CustomerServiceImpl.java
@@ -1,18 +1,19 @@
 package com.example.bankingsystem.service.impl;
 
 import com.example.bankingsystem.converter.CustomerConverter;
-import com.example.bankingsystem.model.dto.request.CustomerCreateRequestDTO;
-import com.example.bankingsystem.model.dto.request.CustomerUpdateRequestDTO;
-import com.example.bankingsystem.model.dto.request.UserRegisterRequest;
-import com.example.bankingsystem.model.dto.response.CustomerGetResponseDTO;
-import com.example.bankingsystem.model.entity.Account;
-import com.example.bankingsystem.model.entity.Customer;
 import com.example.bankingsystem.exception.ServiceOperationAlreadyDeletedException;
 import com.example.bankingsystem.exception.ServiceOperationCanNotCreateException;
 import com.example.bankingsystem.exception.ServiceOperationCanNotDeleteException;
 import com.example.bankingsystem.exception.ServiceOperationNotFoundException;
+import com.example.bankingsystem.model.dto.request.CustomerCreateRequestDTO;
+import com.example.bankingsystem.model.dto.request.CustomerUpdateRequestDTO;
+import com.example.bankingsystem.model.dto.response.CustomerGetResponseDTO;
+import com.example.bankingsystem.model.entity.Account;
+import com.example.bankingsystem.model.entity.Card;
+import com.example.bankingsystem.model.entity.Customer;
 import com.example.bankingsystem.repository.CustomerRepository;
 import com.example.bankingsystem.service.AccountService;
+import com.example.bankingsystem.service.CardService;
 import com.example.bankingsystem.service.CustomerService;
 import lombok.extern.slf4j.Slf4j;
 import org.hibernate.exception.ConstraintViolationException;
@@ -25,17 +26,20 @@ import java.util.Collection;
 import java.util.Objects;
 
 @Service
+
 @Slf4j
 public class CustomerServiceImpl implements CustomerService {
 
     private final CustomerRepository customerRepository;
     private final CustomerConverter customerConverter;
     private final AccountService accountService;
+    private final CardService cardService;
 
-    public CustomerServiceImpl(CustomerRepository customerRepository, CustomerConverter customerConverter, @Lazy AccountService accountService) {
+    public CustomerServiceImpl(CustomerRepository customerRepository, CustomerConverter customerConverter, @Lazy AccountService accountService, @Lazy CardService cardService) {
         this.customerRepository = customerRepository;
         this.customerConverter = customerConverter;
         this.accountService = accountService;
+        this.cardService = cardService;
     }
 
     @Override
@@ -85,23 +89,30 @@ public class CustomerServiceImpl implements CustomerService {
     @Override
     @Transactional
     public String deleteCustomer(Long id, boolean hardDelete) {
+
         Customer customer = customerRepository.getById(id);
 
         Collection<Account> accountCollection = accountService.getAllAccountOneCustomer(id).stream().filter(i -> (i.getBalance().compareTo(BigDecimal.ZERO) > 0)).toList();
-
+        Collection<Card> cardDebtCollection = cardService.getAllCardByCustomerId(id)
+                .stream().filter(i -> i.getCardDebt().compareTo(BigDecimal.ZERO) > 0).toList();
+        Collection<Card> cardBalanceCollection = cardService.getAllCardByCustomerId(id)
+                .stream().filter(i -> i.getCardBalance().compareTo(BigDecimal.ZERO) > 0).toList();
+        
         if (accountCollection.isEmpty()) {
+            if (cardBalanceCollection.isEmpty() && cardDebtCollection.isEmpty()) {
+                if (customer.isDeleted()) {
+                    throw new ServiceOperationAlreadyDeletedException.CustomerAlreadyDeletedException("Customer is already deleted");
+                }
+                if (hardDelete) {
+                    customerRepository.removeCustomerById(id);
+                    return "Customer is deleted successfully";
+                }
 
-            if (customer.isDeleted()) {
-                throw new ServiceOperationAlreadyDeletedException.CustomerAlreadyDeletedException("Customer is already deleted");
-            }
-            if (hardDelete) {
-                customerRepository.removeCustomerById(id);
+                customer.setDeleted(true);
+                customerRepository.save(customer);
                 return "Customer is deleted successfully";
             }
-
-            customer.setDeleted(true);
-            customerRepository.save(customer);
-            return "Customer is deleted successfully";
+            throw new ServiceOperationCanNotDeleteException.CustomerBalanceNotZero("Customer has at least one card which has a debt or balance! ");
         }
         throw new ServiceOperationCanNotDeleteException.CustomerBalanceNotZero("Customer has account that balance of it has money");
     }
@@ -112,7 +123,6 @@ public class CustomerServiceImpl implements CustomerService {
         customerRepository.save(customer);
 
     }
-
 
 
 }
